FLUME-3195. Split the KafkaChannelTest to avoid timeouts
KafkaChannelTest had quite a few test methods so during the build sometimes it
caused a timeout. Refactoring it to smaller test classes decrease the chance of
timeouts.

This closes #183

Reviewers: Miklos Csanady, Ferenc Szabo

(Viktor Somogyi via Ferenc Szabo)
